+++ src/BasicDisplays/ModelDisplay.tsx
@@ -6,6 +6,7 @@ import { cn } from '@/lib/utils';
 export interface ModelDisplayProps extends FormDisplayChildProps {
   className?: string;
   getDisplayValue?: (value: FormDisplayChildProps['value']) => ReactNode;
+  noneValue?: ReactNode;
 }
 
 export const ModelDisplay = forwardRef<HTMLSpanElement, ModelDisplayProps>(
@@ -13,19 +14,21 @@ export const ModelDisplay = forwardRef<HTMLSpanElement, ModelDisplayProps>(
     {
       value,
       getDisplayValue = (val) => {
-        return val?.displayValue ?? 'None';
+        return val?.displayValue;
       },
+      noneValue = 'None',
       className,
       ...props
     },
     ref,
   ) => {
-    const parsedValue =
-      Array.isArray(value) && value.length > 0 ? value : [value];
-    const modelList = parsedValue
-      .flatMap((val: ReactNode) => [getDisplayValue(val), ', '])
-      .slice(0, -1);
-
+    const parsedValue = Array.isArray(value) || !value ? value : [value];
+    const modelList =
+      !parsedValue || parsedValue.length === 0
+        ? noneValue
+        : parsedValue
+            .flatMap((val: any) => [getDisplayValue(val), ', '])
+            .slice(0, -1);
     return (
       <span
         ref={ref}