+++ src/BasicInputs/ModelInput.tsx
@@ -9,20 +9,16 @@ import { SelectInput } from './SelectInput';
 export const ModelInput = forwardRef<
   ElementRef<typeof SelectInput>,
   ComponentPropsWithoutRef<typeof SelectInput>
->(({ value, onChange, options, ...selectInputProps }, ref) => {
-  const isArray = Array.isArray(value);
+>(({ ...selectInputProps }, ref) => {
   return (
     <SelectInput
       ref={ref}
-      value={value}
-      onChange={onChange}
-      getOptionLabel={(option: typeof value) => option.displayValue}
-      getOptionValue={(option: typeof value) => option.id}
-      options={options}
-      isMulti={isArray}
-      clearValue={isArray ? () => [] : undefined}
-      isClearable={!isArray}
-      closeMenuOnSelect={!isArray}
+      getOptionLabel={(option) => option.displayValue}
+      getOptionValue={(option) => option.id}
+      getNewOptionData={(inputValue, optionLabel) => ({
+        displayValue: optionLabel,
+        id: inputValue,
+      })}
       {...selectInputProps}
     />
   );