+++ src/BasicInputs/stories/ModelInput.stories.tsx
@@ -2,24 +2,24 @@ import { useState } from 'react';
 
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { SelectInput } from '../SelectInput';
+import { ModelInput } from '../ModelInput';
 
 const meta = {
-  title: 'Commons/BasicInputs/SelectInput',
-  component: SelectInput,
+  title: 'Commons/BasicInputs/ModelInput',
+  component: ModelInput,
   tags: ['autodocs'],
   args: {
     options: [
-      { value: 'apple', label: 'Apple' },
-      { value: 'mountain', label: 'Mountain' },
-      { value: 'ocean', label: 'Ocean' },
-      { value: 'whisper', label: 'Whisper' },
-      { value: 'butterfly', label: 'Butterfly' },
-      { value: 'galaxy', label: 'Galaxy' },
-      { value: 'harmony', label: 'Harmony' },
-      { value: 'puzzle', label: 'Puzzle' },
-      { value: 'symphony', label: 'Symphony' },
-      { value: 'labyrinth', label: 'Labyrinth' },
+      { id: 'apple', displayValue: 'Apple' },
+      { id: 'mountain', displayValue: 'Mountain' },
+      { id: 'ocean', displayValue: 'Ocean' },
+      { id: 'whisper', displayValue: 'Whisper' },
+      { id: 'butterfly', displayValue: 'Butterfly' },
+      { id: 'galaxy', displayValue: 'Galaxy' },
+      { id: 'harmony', displayValue: 'Harmony' },
+      { id: 'puzzle', displayValue: 'Puzzle' },
+      { id: 'symphony', displayValue: 'Symphony' },
+      { id: 'labyrinth', displayValue: 'Labyrinth' },
     ],
   },
   render: (props) => {
@@ -27,11 +27,11 @@ const meta = {
     return (
       <div className="space-y-2">
         <div>Value: {JSON.stringify(values)}</div>
-        <SelectInput {...props} value={values} onChange={setValues} />
+        <ModelInput {...props} value={values} onChange={setValues} />
       </div>
     );
   },
-} satisfies Meta<typeof SelectInput>;
+} satisfies Meta<typeof ModelInput>;
 export default meta;
 
 type Story = StoryObj<typeof meta>;