+++ src/ModelTable/ModelTable.tsx
@@ -74,7 +74,7 @@ export const ModelTable = Object.assign(
       >
         <BorderWrapper
           bordered={typeof bordered === 'object' ? true : bordered ?? true}
-          className={typeof bordered === 'object' ? bordered?.className : ''}
+          className={cn(typeof bordered === 'object' && bordered?.className)}
         >
           <DnDContextWrapper
             draggable={draggable ?? true}
@@ -87,9 +87,9 @@ export const ModelTable = Object.assign(
               scrollable={
                 typeof scrollable === 'object' ? true : scrollable ?? true
               }
-              className={
-                typeof scrollable === 'object' ? scrollable?.className : ''
-              }
+              className={cn(
+                typeof scrollable === 'object' && scrollable?.className,
+              )}
             >
               <Table
                 ref={ref}