+++ src/ModelTable/ModelTableActionHead.tsx
@@ -1,4 +1,5 @@
 import { TableHead, type TableHeadProps } from '@/Table';
+import { cn } from '@/lib/utils';
 
 import { ACTION_COLUMN } from './ModelTable';
 
@@ -7,10 +8,15 @@ export interface ModelTableActionHeadProps
 
 export const ModelTableActionHead = ({
   children,
+  className,
   ...props
 }: ModelTableActionHeadProps) => {
   return (
-    <TableHead columnId={ACTION_COLUMN} {...props}>
+    <TableHead
+      className={cn('sticky right-0 bg-inherit shadow-left', className)}
+      columnId={ACTION_COLUMN}
+      {...props}
+    >
       {children === undefined ? null : children}
     </TableHead>
   );