+++ src/lib/components/ui/table.tsx
@@ -56,7 +56,7 @@ const TableRow = React.forwardRef<
   <tr
     ref={ref}
     className={cn(
-      'border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted',
+      'border-b bg-background transition-color hover:bg-muted-subtle data-[state=selected]:bg-muted',
       className,
     )}
     {...props}